pyproject.toml
[tool.poetry]
name = "schumacher"
version = "0.1.0"
description = ""
authors = ["Nikita"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datasets = "^2.17.1"
pypdf = "^4.1.0"
torch = "^2.2.1"
openai = "^1.13.3"
pillow = "^10.2.0"
transformers = "^4.38.2"
httpx = "^0.27.0"
vtracer = "^0.6.10"
pypdfium2 = "^4.27.0"
tqdm = "^4.66.2"
sqlalchemy = "^2.0.28"
pytesseract = "^0.3.10"
asynciolimiter = "^1.0.0"
networkx = "^3.2.1"
albumentations = "^1.4.1"
pydantic = "^2.6.3"
pycocotools = "^2.0.7"
torchvision = "^0.17.1"
nltk = "^3.8.1"
datasketch = "^1.6.4"
fastapi = "^0.110.0"
pytorch-lightning = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
hiro = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
